requirements.txt



fastapi
uvicorn
python-multipart
python-docx
reportlab
arabic_reshaper
python-bidi
